+++ src/app/destination/components/activity/activity.component.html
@@ -1,0 +1,11 @@
+<div class="card">
+  <div class="card-image">
+    <img [src]="activity.thumbnail" alt="thumbnail">
+  </div>
+  <div class="card-content">
+    <h2>{{activity.name}}</h2>
+    <p>
+      {{activity.description}}
+    </p>
+  </div>
+</div>
