+++ src/app/destination/components/activity/activity.component.scss
@@ -1,0 +1,14 @@
+.card-image {
+	img {
+		object-fit: cover;
+		width: 100%;
+		height: 100%;
+	}
+	height: 180px;
+	overflow: hidden;
+}
+
+.card {
+	overflow: hidden;
+	height: 350px;
+}
