+++ src/app/destination/components/activity/activity.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit, Input } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
+import { IActivity } from '../../models/activity.model';
+
+@Component({
+  selector: 'app-activity',
+  templateUrl: './activity.component.html',
+  styleUrls: ['./activity.component.scss']
+})
+
+export class ActivityComponent implements OnInit {
+  @Input() activity: IActivity;
+	ngOnInit() {
+	}
+}
