+++ src/app/destination/destination.component.html
@@ -1,39 +1,48 @@
 <main class="main-content" [style.backgroundImage]="bgImg">
 	<div class="container">
 		<div class="grid">
-				<section class="grid-md8 grid-lg7 presentation-section">
-						<nav class="breadcrumb">
-							<a href="#" class="breadcrumb-item">Destinations</a>
-							<span class="breadcrumb-item is-active">{{destination.name}}</span>
-						</nav>
-						<h1>{{destination.name}}</h1>
-						<div class="card">
-							<div class="card-content">
-								{{destination.description}}
-							</div>
-						</div>
-					</section>
-					<section class="grid-md4 grid-lg3 grid-lgOffset1 infos-section">
-						<div class="infos">
+			<section class="grid-md8 grid-lg7 presentation-section">
+				<nav class="breadcrumb">
+					<a href="#" class="breadcrumb-item">Destinations</a>
+					<span class="breadcrumb-item is-active">{{destination.name}}</span>
+				</nav>
+				<h1>{{destination.name}}</h1>
+				<div class="card">
+					<div class="card-content">
+						{{destination.description}}
+					</div>
+				</div>
+				<div class="d-flex">
+					<h1>Activités touristiques</h1>
+					<div *ngFor="let activity of activities">
+						<app-activity [activity]="activity"></app-activity>
+					</div>
+				</div>
+			</section>
+			
+			<section class="grid-md4 grid-lg3 grid-lgOffset1 infos-section">
+				<div class="infos">
+					<ul>
+						<li><b>Pays :</b> {{destination.country}}</li>
+						<li><b>Prix :</b> {{destination.priceRange }}</li>
+						<li><b>Note :</b>
+							<i class="lucca-icon icon-star"></i>
+							<i class="lucca-icon icon-star"></i>
+							<i class="lucca-icon icon-star"></i>
+							<i class="lucca-icon icon-star"></i>
+							<i class="lucca-icon icon-unstared"></i></li>
+						<li>
 							<ul>
-								<li><b>Pays :</b> {{destination.country}}</li>
-								<li><b>Prix :</b> {{destination.priceRange }}</li>
-								<li><b>Note :</b>
-									<i class="lucca-icon icon-star"></i>
-									<i class="lucca-icon icon-star"></i>
-									<i class="lucca-icon icon-star"></i>
-									<i class="lucca-icon icon-star"></i>
-									<i class="lucca-icon icon-unstared"></i></li>
-								<li>
-									<ul>
-										<!-- <b>Tags</b><br> -->
-										<li class="tag" *ngFor="let tag of destination.tags">{{tag}}</li>
-									</ul>
-								</li>
+								<!-- <b>Tags</b><br> -->
+								<li class="tag" *ngFor="let tag of destination.tags">{{tag}}</li>
 							</ul>
-						</div>
-						<app-destination-gallery [destination]="destination"></app-destination-gallery>
-					</section>
+						</li>
+					</ul>
+				</div>
+				<app-destination-gallery [destination]="destination"></app-destination-gallery>
+			</section>
+			
+
 		</div>
 	</div>
 </main>
