+++ src/app/destination/destination.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { IDestination } from './destination.model';
+import { IDestination } from './models/destination.model';
+import { IActivity } from './models/activity.model';
 
 @Component({
 	selector: 'app-destination',
@@ -9,14 +10,16 @@ import { IDestination } from './destination.model';
 })
 export class DestinationComponent implements OnInit {
 	destination: IDestination;
+	activities: IActivity[];
 	get bgImg() { return `url('${this.destination.bg}')`; }
 	constructor(
 		protected route: ActivatedRoute,
 	) {}
 	ngOnInit() {
 		this.route.data
-		.subscribe((data: { destination: IDestination }) => {
+		.subscribe((data: { destination: IDestination, activities: IActivity[] }) => {
 			this.destination = data.destination;
+			this.activities = data.activities;
 		});
 	}
 }
