--- src/app/destination/destination.mock.spec.ts
@@ -1,10 +1,0 @@
-import { IDestination } from './destination.model';
-
-export const paris = {
-	id: 'paris',
-	name: 'paris',
-} as IDestination;
-export const marseille = {
-	id: 'marseille',
-	name: 'marseille',
-} as IDestination;
