+++ src/app/destination/destination.module.ts
@@ -1,12 +1,15 @@
 import { NgModule } from '@angular/core';
-import { DestinationService } from './destination.service';
+import { DestinationService } from './services/destination.service';
 import { DestinationRoutingModule } from './destination.router';
 import { DestinationComponent } from './destination.component';
 import { DestinationThumbnailComponent, DestinationGalleryComponent } from './components';
-import { DestinationResolver } from './destination.resolver';
+import { DestinationResolver } from './resolvers/destination.resolver';
 import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import { ApiModule } from '../api';
+import { ActivityComponent } from './components/activity/activity.component';
+import { ActivityService } from './services/activity.service';
+import { ActivityResolver } from './resolvers/activity.resolver';
 
 @NgModule({
 	imports: [
@@ -18,11 +21,14 @@ import { ApiModule } from '../api';
 	providers: [
 		DestinationService,
 		DestinationResolver,
+		ActivityService,
+		ActivityResolver
 	],
 	declarations: [
 		DestinationComponent,
 		DestinationThumbnailComponent,
 		DestinationGalleryComponent,
+		ActivityComponent,
 	],
 	exports: [
 		DestinationThumbnailComponent,
