+++ src/app/destination/index.ts
@@ -1,3 +1,5 @@
-export * from './destination.model';
+export * from './models/destination.model';
 export * from './destination.module';
-export * from './destination.service';
+export * from './services/destination.service';
+export * from './models/activity.model';
+export * from './services/activity.service';
