+++ src/app/destination/models/activity.model.ts
@@ -1,0 +1,5 @@
+export interface IActivity {
+	description: string;
+	thumbnail: string;
+	destinationId: string;
+}
