+++ src/app/destination/models/destination.model.ts
@@ -1,0 +1,13 @@
+export interface IDestination {
+	id: string;
+	name: string;
+	description: string;
+	country: string;
+	priceRange: string;
+	rating: number;
+	tags: string[];
+	img: string;
+	bg: string;
+	thumbnail: string;
+	gallery: string[];
+}
