+++ src/app/destination/services/activity.mock.spec.ts
@@ -1,0 +1,6 @@
+import { IActivity } from "../models/activity.model";
+
+
+export const marseilleActivity = {
+	destinationId: '4',
+} as IActivity;
