+++ src/app/destination/services/destination.mock.spec.ts
@@ -1,0 +1,10 @@
+import { IDestination } from '../models/destination.model';
+
+export const paris = {
+	id: 'paris',
+	name: 'paris',
+} as IDestination;
+export const marseille = {
+	id: 'marseille',
+	name: 'marseille',
+} as IDestination;
